dom/indexedDB/test/test_objectStore_inline_autoincrement_key_added_on_put.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Property Test</title>

  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="helpers.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript;version=1.7">

function testSteps()
{
  const IDBObjectStore = Components.interfaces.nsIIDBObjectStore;
  const IDBDatabaseException = Components.interfaces.nsIIDBDatabaseException;
  const name = window.location.pathname;
  const description = "My Test Database";

  var request = indexedDB.open(name, description);
  request.onerror = errorHandler;
  request.onsuccess = continueTest;
  var event = yield;

  var db = event.result;

  var test = {
    name: "inline key; key generator",
    autoIncrement: true,
    storedObject: {name: "Lincoln"},
    keyName: "id",
  };

  let request = db.createObjectStore(test.name, test.keyName,
                                     test.autoIncrement);
  request.onerror = errorHandler;
  request.onsuccess = continueTest;
  event = yield;

  request = db.openObjectStore(test.name);
  request.onerror = errorHandler;
  request.onsuccess = continueTest;
  event = yield;

  let objectStore = event.result;
  request = objectStore.put(test.storedObject);
  request.onerror = errorHandler;
  request.onsuccess = continueTest;
  event = yield;

  let id = event.result;
  request = objectStore.get(id);
  request.onerror = errorHandler;
  request.onsuccess = continueTest;
  event = yield;

  // Sanity check!
  SimpleTest.is(event.result.name, test.storedObject.name,
                "The correct object was stored.");

  // Ensure that the id was also stored on the object.
  SimpleTest.is(event.result.id, id, "The object had the id stored on it.");

  SimpleTest.executeSoon(finishTest);
  yield;
}

var gGen = testSteps();
function run_test()
{
  SimpleTest.waitForExplicitFinish();
  continueTest();
}

  </script>

</head>

<body onload="run_test();">
</body>
</html>
